requirements.txt
streamlit==1.30.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
pinecone-client==3.0.0
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
python-multipart==0.0.6